Add files via upload
A custom image classifier that you will train on the fly in the browser using TensorFlow.js, a powerful and flexible machine learning library for Javascript. 

You will first load and run a popular pre-trained model called MobileNet for image classification in the browser. 

You will then use a technique called "transfer learning", which bootstraps our training with the pre-trained MobileNet model and customizes it to train for our application.